requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-appconf==1.0.6
django-environ==0.11.2
django-imagekit==5.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
kavenegar==1.1.2
pilkit==3.0
pillow==11.0.0
promise==2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3

